% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_month_nsa}
\alias{county_month_nsa}
\title{County-Month}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1126122 rows and 16 columns.}
\source{
[https://www.bls.gov/lau/]()
}
\usage{
county_month_nsa
}
\description{
Monthly series (not seasonly adjusted) of employment status of the civilian noninstitutional population from BLS.
}
\keyword{datasets}
