% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_year}
\alias{county_year}
\title{County by month level US employment data}
\format{A data frame with ten variables:
* `laus_code` = "LAUS Code",
* `state_fips` = "State FIPS Code",
* `county_fips` = "County FIPS Code",
* `county` = "County Name",
* `state` = "State Abbreviation",
* `year` = "Year",
* `labor_force` = "Total numer of people in civilian labor force",
* `employed` = "Total number of people employed"
* `unemployed` = "Total number of people unemployed",
* `unem_rate` = "Unemployment rate (= unemployment / labor force; Age: 16 years and over)"}
\source{
[https://www.bls.gov/lau/]()
}
\usage{
county_year
}
\description{
Annual series of US county employment data
}
\keyword{datasets}

