% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_year}
\alias{state_year}
\title{State by year level US employment data}
\format{A data frame with ten variables:
* `fips` = "FIPS code",
* `state` = "State or area",
* `year` = "Year",
* `pop` = "Civilian non-institutional population",
* `clf` = "Total numer of people in civilian labor force",
* `pc_clf` = "Labor force participation rate (= labor force / population; Age: 16 years and over)",
* `emp` = "The number of people employed",
* `pc_emp` = "Employment-population ratio (= employment / population; Age: 16 years and over)",
* `unem` = "The number of people unemployed",
* `unem_rate` = "Unemployment rate (= unemployment / labor force; Age: 16 years and over)"}
\source{
[http://www.bls.gov/lau/rdscnp16.htm]()
}
\usage{
state_year
}
\description{
Annual averages series of employment status of the civilian noninstitutional population from BLS.
}
\keyword{datasets}

